new files:
init.txt --> set the initial parameters of the application such as options or opened files.

debug_info_tasks.txt --> if debug is on, the application will save all the tasks informations (starting and ending times, executing and interrupted tasks etc) in this file. Good for history and tracking tasks.

There's still a lot to do, this is only a partial commit, no complete features has been added yet.